feat: transfer safe_copy function to utils.py file
